+++ src/workerckl.h
@@ -1,7 +1,7 @@
 /*
  * STIGQter - STIG fun with Qt
  *
- * Copyright © 2020–2021 Jon Hood, http://www.hoodsecurity.com/
+ * Copyright © 2021 Jon Hood, http://www.hoodsecurity.com/
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,31 +17,34 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef WORKERASSETCKL_H
-#define WORKERASSETCKL_H
+#ifndef WORKERCKL_H
+#define WORKERCKL_H
 
 #include "asset.h"
+#include "stig.h"
 #include "worker.h"
 
 #include <QObject>
 #include <QXmlStreamWriter>
 
-class WorkerAssetCKL : public Worker
+class WorkerCKL : public Worker
 {
     Q_OBJECT
 
 private:
     QString _fileName;
     Asset _asset;
+    QList<STIG> _stigs;
     void WriteXMLEntry(QXmlStreamWriter &stream, const QString &name, const QString &value);
+    void AddSTIGs(const QVector<STIG> &stigs);
 
 public:
-    explicit WorkerAssetCKL(QObject *parent = nullptr);
-    void AddAsset(const Asset &asset);
+    explicit WorkerCKL(QObject *parent = nullptr);
+    void AddAsset(const Asset &asset, const QVector<STIG> &stigs = {});
     void AddFilename(const QString &name);
 
 public Q_SLOTS:
     void process() override;
 };
 
-#endif // WORKERASSETCKL_H
+#endif // WORKERCKL_H